updating state correctly for configs
